v1/requirements.txt
APScheduler==3.10.4
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
numpy==2.2.4
python-dotenv==1.0.1
sounddevice==0.5.1
soundfile==0.13.1
